Fix memory leaks and use-after-free in window-choose
Collapsing the tree frees the list items, so we cannot take a reference
into that list before calling collapse.  This could be the source of the
crash reported to the mailing list.  This changes the code so we get a
pointer to the session before collapsing the tree.

Additionally I've found two memory leaks in related code:
* old_list contains all items, so use that for cleanup, otherwise
  some items could leak

* list_copy was never freed